requirements.txt
asgiref==3.4.1
Django==3.2.5
django-bootstrap-form==3.4
django-environ==0.4.5
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
